+++ src/statusBar.ts
@@ -1,6 +1,8 @@
 import { StatusBarAlignment, StatusBarItem, window, workspace } from "vscode"; 
 import { Locators } from "../vscode-project-manager-core/src/model/locators";
 import { Project, ProjectStorage } from "../vscode-project-manager-core/src/model/storage";
+import { codicons } from "vscode-ext-codicons";
+import { isRemoteUri } from "../vscode-project-manager-core/src/utils/remote";
 
 let statusItem: StatusBarItem;
 
@@ -10,15 +12,21 @@ export function showStatusBar(projectStorage: ProjectStorage, locators: Locators
 
   // multi-root - decide do use the "first folder" as the original "rootPath"
   // let currentProjectPath = vscode.workspace.rootPath;
-  const workspace0 = workspace.workspaceFolders ? workspace.workspaceFolders[0] : undefined;
-  const currentProjectPath = workspace0 ? workspace0.uri.fsPath : undefined;
+//   const workspace0 = workspace.workspaceFolders ? workspace.workspaceFolders[0] : undefined;
+//   const currentProjectPath = workspace0 ? workspace0.uri.fsPath : undefined;
+  const workspace0 = workspace.workspaceFile ? workspace.workspaceFile :
+                        workspace.workspaceFolders ? workspace.workspaceFolders[0].uri : 
+                        undefined;
+  const currentProjectPath = workspace0 ? workspace0.fsPath : undefined;
 
   if (!showStatusConfig || !currentProjectPath) { return; }
+
+  const isRemote = isRemoteUri(workspace0);
 
   if (!statusItem) {
       statusItem = window.createStatusBarItem(StatusBarAlignment.Left);
   }
-  statusItem.text = "$(file-directory) ";
+  statusItem.text = isRemote ? codicons.remote_explorer + " " : codicons.file_directory + " ";
   statusItem.tooltip = currentProjectPath;
 
   const openInNewWindow: boolean = workspace.getConfiguration("projectManager").get("openInNewWindowWhenClickingInStatusBar", false);
@@ -35,25 +43,26 @@ export function showStatusBar(projectStorage: ProjectStorage, locators: Locators
       return;
   }
 
-//   if (projectStorage.length() === 0) {
-//       return;
-//   }
-
-  let foundProject: Project = projectStorage.existsWithRootPath(currentProjectPath);
-  if (!foundProject) {
-      foundProject = locators.vscLocator.existsWithRootPath(currentProjectPath);
-  }
-  if (!foundProject) {
-      foundProject = locators.gitLocator.existsWithRootPath(currentProjectPath);
-  }
-  if (!foundProject) {
-      foundProject = locators.mercurialLocator.existsWithRootPath(currentProjectPath);
-  }
-  if (!foundProject) {
-      foundProject = locators.svnLocator.existsWithRootPath(currentProjectPath);
-  }
-  if (!foundProject) {
-      foundProject = locators.anyLocator.existsWithRootPath(currentProjectPath);
+  let foundProject: Project;
+  if (isRemote) {
+      foundProject = projectStorage.existsRemoteWithRootPath(workspace0);
+  } else {
+      foundProject = projectStorage.existsWithRootPath(currentProjectPath);
+      if (!foundProject) {
+          foundProject = locators.vscLocator.existsWithRootPath(currentProjectPath);
+      }
+      if (!foundProject) {
+          foundProject = locators.gitLocator.existsWithRootPath(currentProjectPath);
+      }
+      if (!foundProject) {
+          foundProject = locators.mercurialLocator.existsWithRootPath(currentProjectPath);
+      }
+      if (!foundProject) {
+          foundProject = locators.svnLocator.existsWithRootPath(currentProjectPath);
+      }
+      if (!foundProject) {
+          foundProject = locators.anyLocator.existsWithRootPath(currentProjectPath);
+      }
   }
   if (foundProject) {
       statusItem.text += foundProject.name;
@@ -62,7 +71,7 @@ export function showStatusBar(projectStorage: ProjectStorage, locators: Locators
 }
 
 export function updateStatusBar(oldName: string, oldPath: string, newName: string): void {
-  if (statusItem.text === "$(file-directory) " + oldName && statusItem.tooltip === oldPath) {
-      statusItem.text = "$(file-directory) " + newName;
+  if (statusItem.text === codicons.file_directory + " " + oldName && statusItem.tooltip === oldPath) {
+      statusItem.text = codicons.file_directory + " " + newName;
   }
 }