+++ src/whats-new/contentProvider.ts
@@ -1,0 +1,301 @@
+/*---------------------------------------------------------------------------------------------
+*  Copyright (c) Alessandro Fragnani. All rights reserved.
+*  Licensed under the MIT License. See License.md in the project root for license information.
+*--------------------------------------------------------------------------------------------*/
+
+// tslint:disable-next-line:max-line-length
+import { ChangeLogItem, ChangeLogKind, ContentProvider, Header, Image, Sponsor, IssueKind } from "../../vscode-whats-new/src/ContentProvider";
+
+export class WhatsNewProjectManagerContentProvider implements ContentProvider {
+
+    public provideHeader(logoUrl: string): Header {
+        return <Header> {logo: <Image> {src: logoUrl, height: 50, width: 50}, 
+            message: `<b>Project Manager</b> helps you to easily access your <b>projects</b>,
+            no matter where they are located. <i>Don't miss those important projects anymore</i>.
+            <br><br>You can define your own <b>Projects</b> (also called <b>Favorites</b>), or choose 
+            for auto-detect <b>Git</b>, <b>Mercurial</b> or <b>SVN</b> repositories, <b>VSCode</b> 
+            folders or <b>any</b> other folder.`};
+    }
+
+    public provideChangeLog(): ChangeLogItem[] {
+        const changeLog: ChangeLogItem[] = [];
+
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "11.3.0", releaseDate: "September 2020" } });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Support <b>$home</b> and <b>\"~\" (tilde)</b> symbol on projectLocation setting",
+                id: 384,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Support <b>~</b> (tilde) symbol on any path related setting",
+                id: 414,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Support <b>glob</b> patterns in <b>ignoredFolders</b> settings",
+                id: 278,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Czech - Simplified Chinese",
+                id: 412,
+                kind: IssueKind.PR,
+                kudos: "@Amereyeu"
+            }
+        });
+
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "11.2.0", releaseDate: "August 2020" } });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Support string array in Settings UI",
+                id: 410,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Use new <b>remote</b> codicon",
+                id: 396,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Weird icon behaviour when no ThemeIcon is selected",
+                id: 392,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "maxDepthRecursion worried about extra trailing path separator",
+                id: 404,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Localization - Simplified Chinese",
+                id: 403,
+                kind: IssueKind.PR,
+                kudos: "@loniceras"
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Typo in What's New",
+                id: 390,
+                kind: IssueKind.PR,
+                kudos: "@geauxtigers"
+            }
+        });
+
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "11.1.0", releaseDate: "June 2020" } });
+        changeLog.push({
+            kind: ChangeLogKind.CHANGED,
+            detail: {
+                message: "Internal commands can't be customisable",
+                id: 388,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Status bar not working on remotes",
+                id: 379,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "<b>Command Palette</b> showing ”Path does not exists” for remote projects",
+                id: 380,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "<b>Open Folder</b> command in Welcome view not working on Windows",
+                id: 387,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Migrate from TSLint to ESLint",
+                id: 360,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Use <b>vscode-ext-codicons</b> package",
+                id: 386,
+                kind: IssueKind.Issue
+            }
+        });
+
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "11.0.1", releaseDate: "May 2020" } });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Path does not exists for SSH remote project",
+                id: 375,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "<b>Open</b> command in Side Bar's context menu does not work",
+                id: 376,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Path does not exists for Docker remote project on Windows",
+                id: 377,
+                kind: IssueKind.Issue
+            }
+        });
+
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "11.0.0", releaseDate: "May 2020" } });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Support open <b>Remote Projects</b>",
+                id: 345,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "<b>Side bar</b> welcome message",
+                id: 353,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Remove <b>vscode</b> module dependency",
+                id: 369,
+                kind: IssueKind.Issue
+            }
+        });
+
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "10.12.0", releaseDate: "April 2020" } });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Setting to choose how <b>Open in New Window</b> command work on empty windows",
+                id: 188,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Typo in README",
+                id: 348,
+                kind: IssueKind.PR,
+                kudos: "@1st"
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.FIXED,
+            detail: {
+                message: "Typo in README",
+                id: 337,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Support VS Code package split",
+                id: 361,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Support VS Code Extension view Context Menu",
+                id: 327,
+                kind: IssueKind.Issue
+            }
+        });
+
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "10.11.0", releaseDate: "March 2020" } });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Increase <b>Remote Development</b> support",
+                id: 323,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "<b>Add Project to Workspace</b> command in Command Palette",
+                id: 283,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Support <b>ThemeIcon</b>",
+                id: 326,
+                kind: IssueKind.Issue
+            }
+        });
+
+        return changeLog;
+    }
+
+    public provideSponsors(): Sponsor[] {
+        const sponsors: Sponsor[] = [];
+        const sponsorCodeStream: Sponsor = <Sponsor> {
+            title: "Try Codestream",
+            link: "https://sponsorlink.codestream.com/?utm_source=vscmarket&utm_campaign=projectmanager&utm_medium=banner",
+            image: "https://alt-images.codestream.com/codestream_logo_projectmanager.png",
+            width: 35,
+            message: `<p>Discussing code is now as easy as highlighting a block and typing a comment right 
+                      from your IDE. Take the pain out of code reviews and improve code quality.</p>`,
+            extra: 
+                `<a title="Try CodeStream" href="https://sponsorlink.codestream.com/?utm_source=vscmarket&utm_campaign=projectmanager&utm_medium=banner">
+                 Try it free</a>` 
+        };
+        sponsors.push(sponsorCodeStream);
+        return sponsors
+    }
+   
+}